Nuevos métodos en la controladora
Se creó el método para inicializar el árbol, así como el método que devuelve los productos existentes entre todas las cafeterías, también se cambió el tipo de las coordenadas de float a double

+++ Cujae/src/logica/Cafeteria.java
@@ -9,7 +9,7 @@ public class Cafeteria extends Lugar {
 
 
 
-	public Cafeteria(String id, float x, float y, String fotoNorte,
+	public Cafeteria(String id, double x, double y, String fotoNorte,
 			String fotoSur, String fotoEste, String fotoOeste,
 			ArrayList<String> productos) {
 		super(id, x, y, fotoNorte, fotoSur, fotoEste, fotoOeste);
